Scoring/requirements.txt
mysql-connector-python
matplotlib==3.7.5
numpy==1.22.0
pandas==2.0.3
scikit-learn==1.3.2
joblib==1.4.2